==> code/0_DCol.ipynb <==
### SAMPLE DATA FROM KAGGLE DATASET
# https://www.kaggle.com/code/alifiyabatterywala/ukraine-vs-russian-topic-modeling/notebook
# Timeframe: from the 24th Feb to 14th Mar, 2022. Equal time span before and after the demostic ban of Twitter in Russia.
# Sample size: 2%

import pandas as pd 
import os
import csv
import warnings
warnings.simplefilter(action='ignore')
import gc
from pathlib import Path
import math
import glob


# create a filelist for reading .gzip datasets 

# method 1
data_location = os.path.join('kaggle', '*.gzip')
# if re-running this code, please download the kaggle dataset first
# and put this code file together with the data folder renamed as "kaggle"

filenames = glob.glob(data_location)
#print(filenames)


# method 2

csv_collection = []
for dirname, _, filenames in os.walk('kaggle'):
    filenames.sort()
    for filename in filenames:
        data_out = os.path.join('data', '*.csv')
        fullpath = os.path.join(dirname, filename)
        csv_collection.append(fullpath)
        
csv_collection

# method 2 will include 'kaggle/.DS_Store' file in the filelist

# sample 2 % of all tweets in the kaggle dataset

columns = ["userid","username","location","language","tweetid","tweetcreatedts","retweetcount", "hashtags", "text", "usercreatedts", "following", "followers","totaltweets"]

filenames.sort()
dataframe_collection = []

for csvfile in filenames:
    df = pd.read_csv(csvfile, compression='gzip', index_col=0,encoding='utf-8', quoting=csv.QUOTE_ALL)
    df = df[columns]
    # sampling 2%
    df = df.sample(n=math.floor(len(df)/50))
    dataframe_collection.append(df)
    print("{}| {} rows".format(Path(csvfile).name, len(df)))
    
gc.collect()

# combine datasets from different days
df_combined = pd.concat(dataframe_collection, axis=0)
print("Number of all sampled tweets", len(df_combined))

# change timestamp columns to time type
df_combined['tweetcreatedts'] = pd.to_datetime(df_combined['tweetcreatedts'], errors='coerce')
df_combined['usercreatedts'] = pd.to_datetime(df_combined['usercreatedts'], errors='coerce')
df_combined['extractedts'] = pd.to_datetime(df_combined['extractedts'], errors='coerce')

# select equal timeframe bafore and after the ban of twitter in Russia
df1 = df_combined[columns][df_combined['tweetcreatedts'] < "2022-03-14 00:00:00" ]
print("Number of all time filtered tweets", len(df1))
df1.head(3)

# delete data allocation
dataframe_collection = None
del dataframe_collection
gc.collect()

# export dataframe 
df1.to_csv("20220502_resampled_dataset.csv",sep = ",", header = True, encoding = "UTF-8",index=False)
