==> code/3_NCon_N.ipynb <==
### Network construction
# data file: preprocessed_data.pkl
# bipartite network user - tweet
# one-mode user network 
# one-mode tweet network

# other combinations:
# one-mode pro-russian / pro-ukrainian user network   
# one-mode user network before / after Twitter ban in Russia   

# export nodelist and edgelist for network visualisation

import pandas as pd
import networkx as nx
import numpy as np
import csv
from networkx.algorithms import bipartite

data = pd.read_pickle("preprocessed_data.pkl")
data.head()

# Get nodes for bipartite network
users = data["userid"]
tweets = data["textid"]

# Create edge list for bipartite network
columns = ["userid", "textid"]
edges = [(user, textid) for user, textid in data[columns].values]

# save edge list to file
# with open('edges.csv','wb') as file:
#     csv_out=csv.writer(out)
#     csv_out.writerow(['name','num'])
#     for row in data:
#         csv_out.writerow(row)

print(users.nunique())
print(tweets.nunique())

# bipartite network construction
B = nx.Graph()
B.name = "Bipartite user-tweet network"
B.add_nodes_from(users, bipartite=0)
B.add_nodes_from(tweets, bipartite=1)
B.add_edges_from(edges)

nx.info(B)

users_nodes = {n for n, a in B.nodes(data=True) if a["bipartite"] == 0}
tweets_nodes = set(B) - users_nodes

B_users = bipartite.projected_graph(B, users_nodes)
B_users.name = "One-mode user network"
B_tweets = bipartite.projected_graph(B, tweets_nodes)
B_tweets.name = "One-mode tweet network"

print(nx.info(B_users))
print(nx.info(B_tweets))

## Compare two timeframes

# Filter tweets by time period: before and after the tweeter ban 
time_before = data[(data['tweetcreatedts'] < '2022-03-04 23:59:59+00:00')]
time_after = data[(data['tweetcreatedts'] > '2022-03-04 23:59:59+00:00')]
print("rows of timeframe 1 dataset:",len(time_before))
print("rows of timeframe 2 dataset:",len(time_after))

### (1) timeframe1: before

# Get nodes for bipartite network
users_1 = time_before["userid"]
tweets_1 = time_before["textid"]

# Create edge list for bipartite network
columns = ["userid", "textid"]
edges_1 = time_before[columns]
edges_1.to_csv("edgelist_t1.csv",sep = ",", header = True, encoding = "UTF-8",index=False)
# same list-type edgelist for network construction
edges_1 = [(user, textid) for user, textid in time_before[columns].values]

print(users_1.nunique(),"unique users in timeframe 1")
print(tweets_1.nunique(),"unique tweets in timeframe 1")

# construct projected one-mode network for users
# export edgelist for visualisation
B1 = nx.Graph()
B1.name = "Bipartite user-tweet network time 1"
B1.add_nodes_from(users_1, bipartite=0)
B1.add_nodes_from(tweets_1, bipartite=1)
B1.add_edges_from(edges_1)

users_nodes = {n for n, a in B1.nodes(data=True) if a["bipartite"] == 0}
tweets_nodes = set(B1) - users_nodes

B1_users = bipartite.weighted_projected_graph(B1, users_nodes)
B1_users.name = "One-mode user network 1"
B1_tweets = bipartite.weighted_projected_graph(B1, tweets_nodes)
B1_tweets.name = "One-mode tweet network 1"

print(nx.info(B1_users))
print(nx.info(B1_tweets))

# write edgelist of projected network
nx.write_weighted_edgelist(B1_users, "projected_w_user_edgelist_1.csv", delimiter=',', encoding='utf-8')   
# format the edgelist file for Gephi
gephi1_edgelist = pd.read_csv("projected_w_user_edgelist_1.csv",header=None)
gephi1_edgelist.columns=['Source','Target','weight']
gephi1_edgelist.to_csv("projected_w_user_edgelist_1.csv",sep = ",", header = True, encoding = "UTF-8",index=False)


### Eigenvector centrality
https://networkx.org/documentation/stable/reference/algorithms/generated/networkx.algorithms.centrality.eigenvector_centrality.html

eigen_c1 = nx.eigenvector_centrality(B1_users, max_iter=100)
#eigen_c1 = dict(sorted(eigen_c1.items(),key= lambda x:x[1]))
len(eigen_c1)
# covert dic to dataframe 
eigen_c1 = pd.DataFrame(list(eigen_c1.items()),columns = ['userid','Eigen_c']) 
eigen_c1

### (2) timeframe2: After

# Get nodes for bipartite network
users_2 = time_after["userid"]
tweets_2 = time_after["textid"]

# Create edge list for bipartite network
columns = ["userid", "textid"]
edges_2 = time_after[columns]
edges_2.to_csv("edgelist_t2.csv",sep = ",", header = True, encoding = "UTF-8",index=False)
# same list-type edgelist for network construction
edges_2 = [(user, textid) for user, textid in time_after[columns].values]

print(users_2.nunique(),"unique users in timeframe 2")
print(tweets_2.nunique(),"unique tweets in timeframe 2")

# construct projected one-mode network for users
# export edgelist for visualisation
B2 = nx.Graph()
B2.name = "Bipartite user-tweet network time 2"
B2.add_nodes_from(users_2, bipartite=0)
B2.add_nodes_from(tweets_2, bipartite=1)
B2.add_edges_from(edges_2)

users_nodes = {n for n, a in B2.nodes(data=True) if a["bipartite"] == 0}
tweets_nodes = set(B2) - users_nodes

B2_users = bipartite.weighted_projected_graph(B2, users_nodes)
B2_users.name = "One-mode user network 2"
B2_tweets = bipartite.weighted_projected_graph(B2, tweets_nodes)
B2_tweets.name = "One-mode tweet network 2"

print(nx.info(B2_users))
print(nx.info(B2_tweets))

# write edgelist of projected network
nx.write_weighted_edgelist(B2_users, "projected_w_user_edgelist_2.csv", delimiter=',', encoding='utf-8') 
# format the edgelist file for Gephi
gephi2_edgelist = pd.read_csv("projected_w_user_edgelist_2.csv",header=None)
gephi2_edgelist.columns=['Source','Target','weight']
gephi2_edgelist.to_csv("projected_w_user_edgelist_2.csv",sep = ",", header = True, encoding = "UTF-8",index=False)


### Eigenvector centrality

eigen_c2 = nx.eigenvector_centrality(B2_users, max_iter=100)
#engen_c2 = dict(sorted(eigen_c2.items(),key= lambda x:x[1]))
len(engen_c2)
# covert dic to dataframe 
eigen_c2 = pd.DataFrame(list(eigen_c2.items()),columns = ['userid','Eigen_c']) 
#eigen_c2

## Export nodelist

### Mark the nodes appeared in the projected user network

nodelist1 = list(B1_users.nodes)
nodelist2 = list(B2_users.nodes)

# lambda functions marking political orientation if related hashtags were found 
node1_check = lambda node: int(node) in nodelist1 # return true if userid of the nodes in overlap list
node2_check = lambda node: int(node) in nodelist2

data['nodeInG1'] = data['userid'].apply(node1_check)
data['nodeInG2'] = data['userid'].apply(node2_check)

# format node ID for Gephi
data["ID"] = data["userid"].apply(lambda x: int(x))

data.head(3)

### Construct a political orientation index for ID

data['R_i'] = np.where(data['pro_Russian'],1,0)
data['U_i'] = np.where(data['pro_Ukraine'],1,0)
data['N_i'] = np.where(data['Neutral'],1,0)

#data.groupby('userid',as_index=False)['R_i'].sum().sort_values(by=['R_i'],ascending=False)
#data.groupby('userid',as_index=False)['U_i'].sum().sort_values(by=['U_i'],ascending=False)
#data.groupby('userid',as_index=False)['N_i'].sum().sort_values(by=['N_i'],ascending=False)
data["polar_i"] = data["U_i"] - data["R_i"]
polar_index = data.groupby('userid',as_index=False)['polar_i'].sum().sort_values(by=['polar_i'],ascending=False)
# positive score: pro-Ukraine, negative score: pro-Russian
polar_index

### Add Eigenvector and political orientation index to nodelist for Gephi

# select info from pre-processed and labelled dataframe
node_columns = ["userid","username","usercreatedts","following","followers","totaltweets","NodeType","ID"]

nodelist1 = data[data['nodeInG1']][node_columns]
nodelist2 = data[data['nodeInG2']][node_columns]

nodelist1 = pd.merge(nodelist1, polar_index, how = 'left',on='userid')
nodelist2 = pd.merge(nodelist2, polar_index, how = 'left',on='userid')

nodelist1 = pd.merge(nodelist1, eigen_c1, how='left',on="userid")
nodelist2 = pd.merge(nodelist2, eigen_c2, how='left',on="userid")

print(nodelist1["ID"].nunique(),"unique users identified in gephi graph in timeframe 1")
print(nodelist2["ID"].nunique(),"unique users identified in gephi graph in timeframe 2")

nodelist1.to_csv("nodelist1.csv",sep = ",", header = True, encoding = "UTF-8",index=False)
nodelist2.to_csv("nodelist2.csv",sep = ",", header = True, encoding = "UTF-8",index=False)
