==> code/5_bipartite_network_analysis.ipynb <==
import pandas as pd
import networkx as nx

# Get nodes for bipartite network
users = pd.read_csv("nodelist2_poli.csv")
# tweets = pd.read_csv("tweet_nodes.csv")
users.head()

print(users.head())
print(tweets.head())

# drop the first column (old indices)
users = users.drop(['Unnamed: 0'], axis=1)
tweets = tweets.drop(['Unnamed: 0'], axis=1)

# Get edge list for bipartite network
edges = pd.read_csv("edges.csv")

print(users.shape)
print(tweets.shape)
print(edges.shape)

print("User attributes:",users.columns.values)
print("Tweet attributes:",tweets.columns.values)

# Create graph with nodes' attributes
B = nx.Graph()
B.add_nodes_from(users["userid"], bipartite=0, username=users["userid"],location=users["location"],
                 usercreatedts=users["usercreatedts"], following=users["following"],followers=users["followers"])
B.add_nodes_from(tweets["textid"], bipartite=1, text=tweets["text"], hashtags=tweets["hashtags"])
B.add_edges_from(edges.values)

from networkx.algorithms import bipartite

# Create one-mode networks
users_nodes = {n for n, a in B.nodes(data=True) if a["bipartite"] == 0}
tweets_nodes = set(B) - users_nodes

B_users = bipartite.projected_graph(B, users_nodes)
B_tweets = bipartite.projected_graph(B, tweets_nodes)

print(nx.info(B_users))
print(nx.info(B_tweets))

users_degcentrality = nx.degree_centrality(B_users)
tweets_degcentrality = nx.degree_centrality(B_tweets) 

# provide the summary information of two projected networks
print('Projected Network of Insects')
print(f'The network has {len(B_users.nodes)} nodes and {len(B_users.edges)} edges.')

print(f'The network is weighted:\t\t {nx.is_weighted(B_users)}')
print(f'The network is directed:\t\t {nx.is_directed(B_users)}')
print(f'The network is bipartite:\t\t {nx.is_bipartite(B_users)}')
print(f'The network is connected:\t\t {nx.is_connected(B_users)}')

print(f'The isolated node in the network:\t {len(list(nx.isolates(B_users)))}')
B_users_del = B_users.copy()
B_users_del.remove_nodes_from(list(nx.isolates(B_users_del))) # delete the isolated nodes

print('-------------')
print(f'After deleting the isolated nodes, the network has {len(B_users_del.nodes)} nodes and {len(B_users_del.edges)} edges.')
# create a subgraph that is the largest connected component
S_users_del = max(nx.connected_components(B_users_del), key=len)
S_users_del = B_users_del.subgraph(S_users_del)

B_users_del_comp = [len(c) for c in sorted(nx.connected_components(B_users_del), key=len, reverse=True)]
print(f'There are {len(B_users_del_comp)} connected components in the network and the largest one has {max(B_users_del_comp)} nodes.')

print('-------------')
print(f'Density:\t\t\t {nx.density(S_users_del)}')
print(f'Network diameter:\t\t {nx.diameter(S_users_del)}')
print(f'Average shortest path length:\t {nx.average_shortest_path_length(S_users_del)}')
print(f'Average degree:\t\t\t {sum([v for (u,v) in list(nx.degree(S_users_del))]) / len(S_users_del)}')
print('-------------')
degree_insect = dict(nx.degree_centrality(S_users_del))
print(f'Node with the highest degree centrality:\t\t {max(degree_insect, key=degree_insect.get)}')
print(f'Node with the lowest degree centrality:\t\t\t {min(degree_insect, key=degree_insect.get)}')
print('')
betweenness = dict(nx.betweenness_centrality(S_users_del))
print(f'Node with the highest betweenese degree centrality:\t {max(betweenness, key=betweenness.get)}')
print(f'Node with the lowest betweenese degree centrality:\t {min(betweenness, key=betweenness.get)}')
print('')
closeness = dict(nx.closeness_centrality(S_users_del))
print(f'Node with the highest closeness degree centrality:\t {max(closeness, key=closeness.get)}')
print(f'Node with the lowest closeness degree centrality:\t {min(closeness, key=closeness.get)}')
print('')
eigenvector = dict(nx.eigenvector_centrality(S_users_del))
print(f'Node with the highest eigenvector degree centrality:\t {max(eigenvector, key=eigenvector.get)}')
print(f'Node with the lowest eigenvector degree centrality:\t {min(eigenvector, key=eigenvector.get)}')
print('')
load = dict(nx.load_centrality(S_users_del))
print(f'Node with the highest load degree centrality:\t\t {max(load, key=load.get)}')
print(f'Node with the lowest load degree centrality:\t\t {min(load, key=load.get)}')

# provide the summary information of two projected networks
print('Projected Network of Tweets')
print(f'The network has {len(B_tweets.nodes)} nodes and {len(B_tweets.edges)} edges.')

print(f'The network is weighted:\t\t {nx.is_weighted(B_tweets)}')
print(f'The network is directed:\t\t {nx.is_directed(B_tweets)}')
print(f'The network is bipartite:\t\t {nx.is_bipartite(B_tweets)}')
print(f'The network is connected:\t\t {nx.is_connected(B_tweets)}')

print(f'The number of isolated nodes in the network:\t {len(list(nx.isolates(B_tweets)))}')
B_tweets_del = B_tweets.copy()
B_tweets_del.remove_nodes_from(list(nx.isolates(B_tweets_del))) # delete the isolated nodes

print('-------------')
print(f'After deleting the isolated nodes, the network has {len(B_tweets_del.nodes)} nodes and {len(B_tweets_del.edges)} edges.')

B_tweets_del_comp = [len(c) for c in sorted(nx.connected_components(B_tweets_del), key=len, reverse=True)]
print(f'There are {len(B_tweets_del_comp)} connected components in the network and the largest one has {max(B_tweets_del_comp)} nodes.')


# create a subgraph that is the largest connected component
S_tweets_del = max(nx.connected_components(B_tweets_del), key=len)
S_tweets_del = B_tweets_del.subgraph(S_tweets_del)
print('-------------')
print(f'Density:\t\t\t {nx.density(S_tweets_del)}')
print(f'Network diameter:\t\t {nx.diameter(S_tweets_del)}')
print(f'Average shortest path length:\t {nx.average_shortest_path_length(S_tweets_del)}')
print(f'Average degree:\t\t\t {sum([v for (u,v) in list(nx.degree(S_tweets_del))]) / len(S_tweets_del)}')
print('-------------')
degree_insect = dict(nx.degree_centrality(S_tweets_del))
print(f'Node with the highest degree centrality:\t\t {max(degree_insect, key=degree_insect.get)}')
print(f'Node with the lowest degree centrality:\t\t\t {min(degree_insect, key=degree_insect.get)}')
print('')
betweenness = dict(nx.betweenness_centrality(S_tweets_del))
print(f'Node with the highest betweenese degree centrality:\t {max(betweenness, key=betweenness.get)}')
print(f'Node with the lowest betweenese degree centrality:\t {min(betweenness, key=betweenness.get)}')
print('')
closeness = dict(nx.closeness_centrality(S_tweets_del))
print(f'Node with the highest closeness degree centrality:\t {max(closeness, key=closeness.get)}')
print(f'Node with the lowest closeness degree centrality:\t {min(closeness, key=closeness.get)}')
print('')
eigenvector = dict(nx.eigenvector_centrality(S_tweets_del))
print(f'Node with the highest eigenvector degree centrality:\t {max(eigenvector, key=eigenvector.get)}')
print(f'Node with the lowest eigenvector degree centrality:\t {min(eigenvector, key=eigenvector.get)}')
print('')
load = dict(nx.load_centrality(S_tweets_del))
print(f'Node with the highest load degree centrality:\t\t {max(load, key=load.get)}')
print(f'Node with the lowest load degree centrality:\t\t {min(load, key=load.get)}')
